SDR-176 refactor: 테스트 패키지 common -> tool 이름 변경
