+++ be/src/test/java/com/jjikmuk/sikdorak/documentationtest/store/StoreVerifyOrSaveDocumentationTest.java
@@ -7,9 +7,9 @@ import static org.hamcrest.Matchers.equalTo;
 import static org.springframework.restdocs.restassured3.RestAssuredRestDocumentation.document;
 
 import com.jjikmuk.sikdorak.common.ResponseCodeAndMessages;
-import com.jjikmuk.sikdorak.common.mock.WireMockPlaceApiTest;
 import com.jjikmuk.sikdorak.documentationtest.InitDocumentationTest;
 import com.jjikmuk.sikdorak.store.controller.request.StoreVerifyOrSaveRequest;
+import com.jjikmuk.sikdorak.tool.mock.WireMockPlaceApiTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
