+++ be/src/test/java/com/jjikmuk/sikdorak/integration/InitIntegrationTest.java
@@ -1,6 +1,6 @@
 package com.jjikmuk.sikdorak.integration;
 
-import com.jjikmuk.sikdorak.common.DatabaseConfigurator;
+import com.jjikmuk.sikdorak.tool.DatabaseConfigurator;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
