+++ be/src/test/java/com/jjikmuk/sikdorak/tool/DataGenerator.java
@@ -1,4 +1,4 @@
-package com.jjikmuk.sikdorak.common;
+package com.jjikmuk.sikdorak.tool;
 
 import com.jjikmuk.sikdorak.comment.domain.Comment;
 import com.jjikmuk.sikdorak.comment.repository.CommentRepository;
